Remove outdated tuning script and enhance plot utility
