+++ scripts/prompt_tuning_plot.py
@@ -4,11 +4,58 @@
 import argparse
 import json
 from pathlib import Path
+from difflib import SequenceMatcher
 import matplotlib
 matplotlib.use("Agg")
 import matplotlib.pyplot as plt
 
-from prompt_tuning import tune_prompt
+import prompt_to_text
+from project_config import PROMPT_CONFIG
+
+
+def match_percent(a: str, b: str) -> float:
+    """Return percent similarity between two strings."""
+    return SequenceMatcher(None, a.lower(), b.lower()).ratio() * 100
+
+
+def refine_prompt(current_prompt: str, input_text: str, output_text: str, target: str) -> str:
+    """Use DeepSeek to suggest a better prompt."""
+    instruction = (
+        f"{PROMPT_CONFIG}Vi prøver å løse teksten '{input_text}'. "
+        f"Den gjeldende prompten er: '{current_prompt}'. "
+        f"Svaret etter forrige denne prompten ble: '{output_text}'. "
+        f"Målet er at svaret skal bli: '{target}'. "
+        "Oppdater prompten som på nytt skal bli sendt til den samme teksten for å komme nærmere svaret. "
+        "Svar kun med selve prompten."
+    )
+    suggestion = prompt_to_text.prompt_to_text(
+        instruction, max_tokens=200, isNum=False, maxLen=1500
+    )
+    return suggestion if suggestion else current_prompt
+
+
+def tune_prompt(initial_prompt: str, input_text: str, target: str, iterations: int = 15):
+    prompts = [initial_prompt]
+    outputs = []
+    matches = []
+    current_prompt = initial_prompt
+
+    for i in range(iterations):
+        print(f"\n--- Iterasjon {i + 1} ---")
+        full_prompt = f"{PROMPT_CONFIG}{current_prompt}\n\n{input_text}"
+        out = prompt_to_text.prompt_to_text(
+            full_prompt, max_tokens=800, isNum=False, maxLen=1500
+        )
+        out = out or ""
+        outputs.append(out)
+        match = match_percent(out, target)
+        matches.append(match)
+        print(f"Match: {match:.2f}%\nOutput: {out}\n")
+        current_prompt = refine_prompt(current_prompt, input_text, out, target)
+        prompts.append(current_prompt)
+        print(f"Prompt etter iterasjon {i + 1}: {current_prompt}")
+
+    return prompts, outputs, matches
 
 
 def save_results(prompts, outputs, matches, path):
